@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.16.1 <3.0.0"
+  sdk: ">=2.19.0 <4.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -35,22 +35,24 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   animated_text_kit: ^4.2.2
-  image_picker: ^0.8.5+3
-  firebase_auth: ^3.6.2
-  cloud_firestore: ^3.4.3
-  firebase_storage: ^10.3.4
-  firebase_core: ^1.20.0
-  uuid: ^3.0.6
+  image_picker: ^1.1.2
+  firebase_auth: ^5.2.1
+  cloud_firestore: ^5.4.1
+  firebase_storage: ^12.3.0
+  firebase_core: ^3.4.1
+  uuid: ^4.5.0
+  intl: ^0.19.0
+   
   staggered_grid_view_flutter: ^0.0.4
   flutter_swiper_null_safety: ^1.0.2
   provider: ^6.0.3
   font_awesome_flutter: ^10.1.0
-  badges: ^2.0.3
+  badges: ^3.1.2
   sn_progress_dialog: ^1.0.8
-  flutter_date_pickers: ^0.4.0
-  flutter_datetime_picker: ^1.5.1
-  flutter_stripe: ^4.0.0
-  http: ^0.13.5
+  flutter_date_pickers: ^0.4.3
+  flutter_datetime_picker_plus: ^2.1.0
+  flutter_stripe: ^11.0.0
+  http: ^1.2.2
   flutter_rating_bar: ^4.0.1
   expandable: ^5.0.1
 
@@ -63,7 +65,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^1.0.0
+  flutter_lints: ^4.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
@@ -120,4 +122,6 @@ flutter:
     - family: Acme
       fonts:
         - asset: fonts/Acme/Acme-Regular.ttf
+
+
 
