@@ -7,9 +7,10 @@ import 'package:multi_store_app/main_screen/dashboard.dart';
 import 'package:multi_store_app/main_screen/home.dart';
 import 'package:multi_store_app/main_screen/store.dart';
 import 'package:multi_store_app/main_screen/upload_product.dart';
+import 'package:badges/badges.dart' as badges;
 
 class SupplierHomeScreen extends StatefulWidget {
-  const SupplierHomeScreen({Key? key}) : super(key: key);
+  const SupplierHomeScreen({super.key});
 
   @override
   State<SupplierHomeScreen> createState() => _SupplierHomeScreenState();
@@ -47,7 +48,8 @@ class _SupplierHomeScreenState extends State<SupplierHomeScreen> {
                 elevation: 0,
                 type: BottomNavigationBarType.fixed,
                 selectedItemColor: Colors.black,
-                selectedLabelStyle: const TextStyle(fontWeight: FontWeight.w600),
+                selectedLabelStyle:
+                    const TextStyle(fontWeight: FontWeight.w600),
                 currentIndex: _selectedIndex,
                 onTap: (index) {
                   setState(() {
@@ -55,22 +57,28 @@ class _SupplierHomeScreenState extends State<SupplierHomeScreen> {
                   });
                 },
                 items: [
-                  const BottomNavigationBarItem(icon: Icon(Icons.home), label: "Home"),
-                  const BottomNavigationBarItem(icon: Icon(Icons.search), label: "Category"),
-                  const BottomNavigationBarItem(icon: Icon(Icons.shop), label: "Store"),
+                  const BottomNavigationBarItem(
+                      icon: Icon(Icons.home), label: "Home"),
+                  const BottomNavigationBarItem(
+                      icon: Icon(Icons.search), label: "Category"),
+                  const BottomNavigationBarItem(
+                      icon: Icon(Icons.shop), label: "Store"),
                   BottomNavigationBarItem(
-                      icon: Badge(
+                      icon: badges.Badge(
                           showBadge: snapshot.data!.docs.isEmpty ? false : true,
-                          padding: const EdgeInsets.all(2),
-                          badgeColor: Colors.yellow,
+                          badgeStyle: const BadgeStyle(
+                            badgeColor: Colors.yellow,
+                          ),
                           badgeContent: Text(
                             snapshot.data!.docs.length.toString(),
-                            style: const TextStyle(fontSize: 18, fontWeight: FontWeight.w600),
+                            style: const TextStyle(
+                                fontSize: 18, fontWeight: FontWeight.w600),
                           ),
-                          animationType: BadgeAnimationType.scale,
+                          badgeAnimation: const BadgeAnimation.scale(),
                           child: const Icon(Icons.dashboard)),
                       label: "DashBoard"),
-                  const BottomNavigationBarItem(icon: Icon(Icons.upload), label: "Upload"),
+                  const BottomNavigationBarItem(
+                      icon: Icon(Icons.upload), label: "Upload"),
                 ]),
           );
         });
