@@ -7,14 +7,16 @@ import 'package:flutter/material.dart';
 import 'package:multi_store_app/widgets/yellow_button.dart';
 
 class WelcomeScreen extends StatefulWidget {
-  const WelcomeScreen({Key? key}) : super(key: key);
+  const WelcomeScreen({super.key});
 
   @override
   State<WelcomeScreen> createState() => _WelcomeScreenState();
 }
 
-class _WelcomeScreenState extends State<WelcomeScreen> with SingleTickerProviderStateMixin {
-  CollectionReference anonymous = FirebaseFirestore.instance.collection('anonymous');
+class _WelcomeScreenState extends State<WelcomeScreen>
+    with SingleTickerProviderStateMixin {
+  CollectionReference anonymous =
+      FirebaseFirestore.instance.collection('anonymous');
   late AnimationController _controller;
   late String _uid;
   List<Color> textColors = [
@@ -47,180 +49,200 @@ class _WelcomeScreenState extends State<WelcomeScreen> with SingleTickerProvider
     return Scaffold(
       body: Container(
         decoration: const BoxDecoration(
-          image: DecorationImage(image: AssetImage("images/inapp/bgimage.jpg"), fit: BoxFit.cover),
+          image: DecorationImage(
+              image: AssetImage("images/inapp/bgimage.jpg"), fit: BoxFit.cover),
         ),
         constraints: const BoxConstraints.expand(),
         child: SafeArea(
-          child: Column(mainAxisAlignment: MainAxisAlignment.spaceBetween, children: [
-            AnimatedTextKit(
-              animatedTexts: [
-                ColorizeAnimatedText("WELCOME",
-                    colors: textColors,
-                    textStyle: const TextStyle(
-                        fontSize: 50, fontWeight: FontWeight.bold, fontFamily: "Acme")),
-                ColorizeAnimatedText("DUCKS STORE",
-                    colors: textColors,
-                    textStyle: const TextStyle(
-                        fontSize: 50, fontWeight: FontWeight.bold, fontFamily: "Acme"))
-              ],
-              isRepeatingAnimation: true,
-              repeatForever: true,
-            ),
-            // const Text("Welcome",
-            //     style: TextStyle(fontSize: 30, fontWeight: FontWeight.bold, color: Colors.white)),
-            const SizedBox(
-              height: 120,
-              width: 200,
-              child: Image(image: AssetImage("images/inapp/logo.jpg")),
-            ),
-            const Text("Shop",
-                style: TextStyle(fontSize: 30, fontWeight: FontWeight.bold, color: Colors.white)),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.end,
+          child: Column(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
               children: [
-                Column(
-                  crossAxisAlignment: CrossAxisAlignment.end,
+                AnimatedTextKit(
+                  animatedTexts: [
+                    ColorizeAnimatedText("WELCOME",
+                        colors: textColors,
+                        textStyle: const TextStyle(
+                            fontSize: 50,
+                            fontWeight: FontWeight.bold,
+                            fontFamily: "Acme")),
+                    ColorizeAnimatedText("DUCKS STORE",
+                        colors: textColors,
+                        textStyle: const TextStyle(
+                            fontSize: 50,
+                            fontWeight: FontWeight.bold,
+                            fontFamily: "Acme"))
+                  ],
+                  isRepeatingAnimation: true,
+                  repeatForever: true,
+                ),
+                // const Text("Welcome",
+                //     style: TextStyle(fontSize: 30, fontWeight: FontWeight.bold, color: Colors.white)),
+                const SizedBox(
+                  height: 120,
+                  width: 200,
+                  child: Image(image: AssetImage("images/inapp/logo.jpg")),
+                ),
+                const Text("Shop",
+                    style: TextStyle(
+                        fontSize: 30,
+                        fontWeight: FontWeight.bold,
+                        color: Colors.white)),
+                Row(
+                  mainAxisAlignment: MainAxisAlignment.end,
                   children: [
-                    Container(
-                      decoration: const BoxDecoration(
-                          color: Colors.white38,
-                          borderRadius: BorderRadius.only(
-                              topLeft: Radius.circular(50), bottomLeft: Radius.circular(50))),
-                      child: const Text("Suppliers only",
-                          style: TextStyle(
-                              fontSize: 30,
-                              color: Colors.yellowAccent,
-                              fontWeight: FontWeight.w600)),
+                    Column(
+                      crossAxisAlignment: CrossAxisAlignment.end,
+                      children: [
+                        Container(
+                          decoration: const BoxDecoration(
+                              color: Colors.white38,
+                              borderRadius: BorderRadius.only(
+                                  topLeft: Radius.circular(50),
+                                  bottomLeft: Radius.circular(50))),
+                          child: const Text("Suppliers only",
+                              style: TextStyle(
+                                  fontSize: 30,
+                                  color: Colors.yellowAccent,
+                                  fontWeight: FontWeight.w600)),
+                        ),
+                        const SizedBox(height: 6),
+                        Container(
+                          height: 60,
+                          width: MediaQuery.of(context).size.width * 0.9,
+                          decoration: const BoxDecoration(
+                            color: Colors.white38,
+                            borderRadius: BorderRadius.only(
+                              topLeft: Radius.circular(50),
+                              bottomLeft: Radius.circular(50),
+                            ),
+                          ),
+                          child: Row(
+                            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                            children: [
+                              AnimatedLogo(controller: _controller),
+                              YellowButton(
+                                  label: "log In",
+                                  onPressed: () {
+                                    Navigator.pushReplacementNamed(
+                                        context, "/supplier_login");
+                                  },
+                                  width: 0.25),
+                              Padding(
+                                padding: const EdgeInsets.only(right: 8.0),
+                                child: YellowButton(
+                                    label: "Sign Up",
+                                    onPressed: () {
+                                      Navigator.pushReplacementNamed(
+                                          context, "/supplier_signup");
+                                    },
+                                    width: 0.25),
+                              ),
+                            ],
+                          ),
+                        ),
+                      ],
                     ),
-                    const SizedBox(height: 6),
+                  ],
+                ),
+                Row(
+                  mainAxisAlignment: MainAxisAlignment.start,
+                  children: [
                     Container(
                       height: 60,
                       width: MediaQuery.of(context).size.width * 0.9,
                       decoration: const BoxDecoration(
                         color: Colors.white38,
                         borderRadius: BorderRadius.only(
-                          topLeft: Radius.circular(50),
-                          bottomLeft: Radius.circular(50),
+                          topRight: Radius.circular(50),
+                          bottomRight: Radius.circular(50),
                         ),
                       ),
                       child: Row(
                         mainAxisAlignment: MainAxisAlignment.spaceBetween,
                         children: [
-                          AnimatedLogo(controller: _controller),
-                          YellowButton(
-                              label: "log In",
-                              onPressed: () {
-                                Navigator.pushReplacementNamed(context, "/supplier_login");
-                              },
-                              width: 0.25),
                           Padding(
-                            padding: const EdgeInsets.only(right: 8.0),
+                            padding: const EdgeInsets.only(left: 8.0),
                             child: YellowButton(
-                                label: "Sign Up",
+                                label: "log In",
                                 onPressed: () {
-                                  Navigator.pushReplacementNamed(context, "/supplier_signup");
+                                  Navigator.pushReplacementNamed(
+                                      context, "/customer_screen");
                                 },
                                 width: 0.25),
                           ),
+                          YellowButton(
+                              label: "Sign Up",
+                              onPressed: () {
+                                Navigator.pushReplacementNamed(
+                                    context, "/customer_signup");
+                              },
+                              width: 0.25),
+                          AnimatedLogo(controller: _controller),
                         ],
                       ),
                     ),
                   ],
                 ),
-              ],
-            ),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.start,
-              children: [
-                Container(
-                  height: 60,
-                  width: MediaQuery.of(context).size.width * 0.9,
-                  decoration: const BoxDecoration(
-                    color: Colors.white38,
-                    borderRadius: BorderRadius.only(
-                      topRight: Radius.circular(50),
-                      bottomRight: Radius.circular(50),
-                    ),
-                  ),
-                  child: Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                    children: [
-                      Padding(
-                        padding: const EdgeInsets.only(left: 8.0),
-                        child: YellowButton(
-                            label: "log In",
-                            onPressed: () {
-                              Navigator.pushReplacementNamed(context, "/customer_screen");
-                            },
-                            width: 0.25),
-                      ),
-                      YellowButton(
-                          label: "Sign Up",
-                          onPressed: () {
-                            Navigator.pushReplacementNamed(context, "/customer_signup");
-                          },
-                          width: 0.25),
-                      AnimatedLogo(controller: _controller),
-                    ],
-                  ),
-                ),
-              ],
-            ),
-            Padding(
-              padding: const EdgeInsets.symmetric(vertical: 8.0),
-              child: Container(
-                  height: 90,
-                  decoration: const BoxDecoration(color: Colors.white38),
-                  child: Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceAround,
-                    children: [
-                      GoogleFacebookLogIn(
-                        label: "Google",
-                        onPressed: () {},
-                        child: const Image(image: AssetImage("images/inapp/google.jpg")),
-                      ),
-                      GoogleFacebookLogIn(
-                        label: "Facebook",
-                        onPressed: () {},
-                        child: const Image(image: AssetImage("images/inapp/facebook.jpg")),
-                      ),
-                      pressing == true
-                          ? const CircularProgressIndicator()
-                          : GoogleFacebookLogIn(
-                              label: "Google",
-                              onPressed: () async {
-                                setState(() {
-                                  pressing = true;
-                                });
-                                await FirebaseAuth.instance
-                                    .signInAnonymously()
-                                    .whenComplete(() async {
-                                  _uid = FirebaseAuth.instance.currentUser!.uid;
-                                  await anonymous.doc(_uid).set({
-                                    'storename': '',
-                                    'name': '',
-                                    'email': '',
-                                    'storelogo': '',
-                                    'coverimage': '',
-                                    'profileImage': '',
-                                    'phone': '',
-                                    'address': '',
-                                    'cid': _uid
-                                  });
-                                });
-
-                                Navigator.pushReplacementNamed(context, '/customer_screen');
-                              },
-                              child: const Icon(
-                                Icons.person,
-                                size: 50,
-                                color: Colors.lightBlueAccent,
-                              ),
+                Padding(
+                  padding: const EdgeInsets.symmetric(vertical: 8.0),
+                  child: Container(
+                      height: 90,
+                      decoration: const BoxDecoration(color: Colors.white38),
+                      child: Row(
+                        mainAxisAlignment: MainAxisAlignment.spaceAround,
+                        children: [
+                          GoogleFacebookLogIn(
+                            label: "Google",
+                            onPressed: () {},
+                            child: const Image(
+                              image: AssetImage("images/inapp/google.jpg"),
                             ),
-                    ],
-                  )),
-            )
-          ]),
+                          ),
+                          GoogleFacebookLogIn(
+                            label: "Facebook",
+                            onPressed: () {},
+                            child: const Image(
+                                image: AssetImage("images/inapp/facebook.jpg")),
+                          ),
+                          pressing == true
+                              ? const CircularProgressIndicator()
+                              : GoogleFacebookLogIn(
+                                  label: "Google",
+                                  onPressed: () async {
+                                    setState(() {
+                                      pressing = true;
+                                    });
+                                    await FirebaseAuth.instance
+                                        .signInAnonymously()
+                                        .whenComplete(() async {
+                                      _uid = FirebaseAuth
+                                          .instance.currentUser!.uid;
+                                      await anonymous.doc(_uid).set({
+                                        'storename': '',
+                                        'name': '',
+                                        'email': '',
+                                        'storelogo': '',
+                                        'coverimage': '',
+                                        'profileImage': '',
+                                        'phone': '',
+                                        'address': '',
+                                        'cid': _uid
+                                      });
+                                    });
+
+                                    Navigator.pushReplacementNamed(
+                                        context, '/customer_screen');
+                                  },
+                                  child: const Icon(
+                                    Icons.person,
+                                    size: 50,
+                                    color: Colors.lightBlueAccent,
+                                  ),
+                                ),
+                        ],
+                      )),
+                )
+              ]),
         ),
       ),
     );
@@ -229,10 +251,9 @@ class _WelcomeScreenState extends State<WelcomeScreen> with SingleTickerProvider
 
 class AnimatedLogo extends StatelessWidget {
   const AnimatedLogo({
-    Key? key,
+    super.key,
     required AnimationController controller,
-  })  : _controller = controller,
-        super(key: key);
+  }) : _controller = controller;
 
   final AnimationController _controller;
 
@@ -253,11 +274,11 @@ class GoogleFacebookLogIn extends StatelessWidget {
   final Function() onPressed;
   final Widget child;
   const GoogleFacebookLogIn({
-    Key? key,
+    super.key,
     required this.label,
     required this.onPressed,
     required this.child,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
